requirements.txt
Django==5.0.6
djangorestframework==3.15.1
pyrebase==3.0.27
python-decouple==3.8
